Matrix_smart in a separate file - new allocate and resize functions
