+++ examples/grayscale_rgba_example.html
@@ -22,7 +22,8 @@
         const U8_t = jsfeat.Types.U8_t.value;
         const C1_t = jsfeat.Types.C1_t.value;
 
-        var m_u8 = new jsfeat.Matrix_smart(img.width, img.height, U8_t | C1_t);
+        var m_u8 = new jsfeat.matrix_smart(img.width, img.height, U8_t | C1_t);
+        //m_u8.allocate()
         console.log(m_u8);
         var imgproc = new jsfeat.imgproc();
