+++ src/jsfeat.h
@@ -1,6 +1,7 @@
 #include <imgproc/imgproc.h>
 #include <keypoint_t/keypoint_t.h>
 #include <keypoints/keypoints.h>
+#include <matrix_smart/matrix_smart.h>
 #include <matrix_t/matrix_t.h>
 #include <orb/orb.h>
 #include <pyramid_t/pyramid_t.h>
